requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.28.0
